+++ src/main/java/com/authlete/jaxrs/api/AuthleteApiImpl.java
@@ -62,6 +62,8 @@ import com.authlete.common.dto.DeviceVerificationRequest;
 import com.authlete.common.dto.DeviceVerificationResponse;
 import com.authlete.common.dto.FederationConfigurationRequest;
 import com.authlete.common.dto.FederationConfigurationResponse;
+import com.authlete.common.dto.FederationRegistrationRequest;
+import com.authlete.common.dto.FederationRegistrationResponse;
 import com.authlete.common.dto.GMRequest;
 import com.authlete.common.dto.GMResponse;
 import com.authlete.common.dto.GrantedScopesGetResponse;
@@ -169,6 +171,7 @@ public class AuthleteApiImpl extends AuthleteApiJaxrsImpl
     private static final String GM_API_PATH                                   = "/api/gm";
     private static final String CLIENT_LOCK_FLAG_UPDATE_API_PATH              = "/api/client/lock_flag/update/%s";
     private static final String FEDERATION_CONFIGURATION_API_PATH             = "/api/federation/configuration";
+    private static final String FEDERATION_REGISTRATION_API_PATH              = "/api/federation/registration";
 
 
     private final String mServiceOwnerAuth;
@@ -1452,5 +1455,16 @@ public class AuthleteApiImpl extends AuthleteApiJaxrsImpl
                 new ServicePostApiCaller<FederationConfigurationResponse>(
                         FederationConfigurationResponse.class, request,
                         FEDERATION_CONFIGURATION_API_PATH));
+    }
+
+
+    @Override
+    public FederationRegistrationResponse federationRegistration(
+            FederationRegistrationRequest request) throws AuthleteApiException
+    {
+        return executeApiCall(
+                new ServicePostApiCaller<FederationRegistrationResponse>(
+                        FederationRegistrationResponse.class, request,
+                        FEDERATION_REGISTRATION_API_PATH));
     }
 }