+++ src/main/java/com/authlete/jaxrs/api/AuthleteApiImplV3.java
@@ -62,6 +62,8 @@ import com.authlete.common.dto.DeviceVerificationRequest;
 import com.authlete.common.dto.DeviceVerificationResponse;
 import com.authlete.common.dto.FederationConfigurationRequest;
 import com.authlete.common.dto.FederationConfigurationResponse;
+import com.authlete.common.dto.FederationRegistrationRequest;
+import com.authlete.common.dto.FederationRegistrationResponse;
 import com.authlete.common.dto.GMRequest;
 import com.authlete.common.dto.GMResponse;
 import com.authlete.common.dto.GrantedScopesGetResponse;
@@ -168,6 +170,7 @@ public class AuthleteApiImplV3 extends AuthleteApiJaxrsImpl
     private static final String GM_API_PATH                                   = "/api/%d/gm";
     private static final String CLIENT_LOCK_FLAG_UPDATE_API_PATH              = "/api/%d/client/lock_flag/update/%s";
     private static final String FEDERATION_CONFIGURATION_API_PATH             = "/api/%d/federation/configuration";
+    private static final String FEDERATION_REGISTRATION_API_PATH              = "/api/%d/federation/registration";
 
 
     private final String mAuth;
@@ -1349,5 +1352,16 @@ public class AuthleteApiImplV3 extends AuthleteApiJaxrsImpl
                 new PostApiCaller<FederationConfigurationResponse>(
                         FederationConfigurationResponse.class, request,
                         FEDERATION_CONFIGURATION_API_PATH, mServiceId));
+    }
+
+
+    @Override
+    public FederationRegistrationResponse federationRegistration(
+            FederationRegistrationRequest request) throws AuthleteApiException
+    {
+        return executeApiCall(
+                new PostApiCaller<FederationRegistrationResponse>(
+                        FederationRegistrationResponse.class, request,
+                        FEDERATION_REGISTRATION_API_PATH, mServiceId));
     }
 }